requirements.txt
-i https://pypi.org/simple
asgiref==3.2.10
attrs==19.3.0
automat==20.2.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
constantly==15.1.0
cryptography==3.0
cssselect==1.1.0
django==3.1
flask-pymongo==2.3.0
flask==1.1.2
google-api-core[grpc]==1.22.1
google-auth==1.20.1
google-cloud-vision==1.0.0
google-cloud==0.34.0
googleapis-common-protos==1.52.0
grpcio==1.31.0
hyperlink==20.0.1
idna==2.10
image==1.5.32
incremental==17.5.0
itemadapter==0.1.0
itemloaders==1.0.2
itsdangerous==1.1.0
jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
lxml==4.5.2
markupsafe==1.1.1
numpy==1.19.1
pandas==1.1.0
parsel==1.6.0
pillow==7.2.0
protego==0.1.16
protobuf==3.12.4
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20
pydispatcher==2.0.5
pyhamcrest==2.0.2
pymongo==3.11.0
pyopenssl==19.1.0
python-dateutil==2.8.1
pytz==2020.1
queuelib==1.5.0
requests==2.24.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.2
scrapy-selenium==0.0.7
scrapy==2.3.0
selenium==3.141.0
service-identity==18.1.0
six==1.15.0
sklearn==0.0
sqlparse==0.3.1
threadpoolctl==2.1.0
twisted==20.3.0
urllib3==1.25.10
w3lib==1.22.0
werkzeug==1.0.1
xlrd==1.2.0
zope.interface==5.1.0